1.重构 命令处理 模块 2.重构 网络处理 模块 利用回调函数，将两个模块彻底解耦合。
