+++ RemoteCtrl/RemoteCtrl/Command.h
@@ -1,13 +1,17 @@
 #pragma once
-#include <map>
-#include "ServerSocket.h"
+
 #include <atlimage.h>
 #include <direct.h>
-#include "Tool.h"
+
+#include <map>
+
 #include <io.h>
 #include <list>
 #include "LockInfoDialog.h"
 #include "resource.h"
+
+#include "Tool.h"
+#include "Packet.h"
 
 class CCommand
 {
@@ -17,31 +21,47 @@ public:
 	~CCommand(){
 	}
 
-	int ExcuteCommand(int nCmd);
+    static void RunCommand(void* arg, int status, std::list<CPacket>& listPackets, CPacket& inPacket) {
+        CCommand* thiz = (CCommand*)arg;
+        if (status > 0) {
+            int ret = thiz->ExcuteCommand(status, listPackets, inPacket);
+            if (ret != 0) {
+                TRACE("[服务器]执行命令失败，%d ret=%d\r\n", status, ret);
+            }
+        }
+        else {
+            MessageBox(NULL, _T("无法正常接入用户，自动重试！"), _T("接入用户失败"), MB_OK | MB_ICONERROR);
+        }
+    }
+
+
+    int ExcuteCommand(int nCmd, std::list<CPacket>& listPackets, CPacket& inPacket);
 
 protected:
-	typedef int(CCommand::* CMDFUNC)();	//成员函数指针
-	std::map<int, CMDFUNC> m_mapFunction;
+
+	typedef int(CCommand::* CMDFUNC)(std::list<CPacket>& listPackets, CPacket& inPacket);	//成员函数指针
+	
+    std::map<int, CMDFUNC> m_mapFunction;
     CLockInfoDialog dlg;
 
     unsigned threadId;
 
+
 protected:
+
     static unsigned __stdcall threadEntryForLockDlg(void* arg)
     {   
         CCommand* thiz = (CCommand*)arg;
         thiz->threadForLockDlg();
         _endthreadex(0);
-
         return 0;
     }
 
     void threadForLockDlg() {
-
         TRACE("[服务器]%s(%d):%d\r\n", __FUNCTION__, __LINE__, GetCurrentThreadId());
         //非模态
         dlg.Create(IDD_DIALOG_INFO, NULL);
-
+         
         dlg.ShowWindow(SW_SHOW);
 
         //全屏遮蔽
@@ -106,7 +126,7 @@ protected:
         dlg.DestroyWindow();
     }
 
-    int MakeDriverInfo() {
+    int MakeDriverInfo(std::list<CPacket>& listPackets, CPacket& inPacket) {
         std::string result;
         for (int i = 1; i <= 26; i++) {
             if (_chdrive(i) == 0) {
@@ -118,27 +138,21 @@ protected:
             }
         }
 
-        CPacket pack(1, (BYTE*)result.c_str(), result.size());
-        CTool::Dump((BYTE*)pack.Data(), pack.Size());
+        listPackets.push_back(CPacket(1, (BYTE*)result.c_str(), result.size()));
 
-        CServerSocket::getInstance()->Send(pack);
         return 0;
     }
 
-    int MakeDirectoryInfo() {
-        std::string strPath;
-        if (CServerSocket::getInstance()->GetFilePath(strPath) == false) {
-            OutputDebugString(_T("当前的命令不是获取文件列表，命令解析错误！"));
-            return -1;
-        }
+    int MakeDirectoryInfo(std::list<CPacket>& listPackets, CPacket& inPacket) {
+
+        std::string strPath = inPacket.strData;
 
         if (_chdir(strPath.c_str()) != 0) {
             FILEINFO finfo;
             finfo.HasNext = FALSE;
-            CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-
-            CServerSocket::getInstance()->Send(pack);
-
+            
+            listPackets.push_back(CPacket(2, (BYTE*)&finfo, sizeof(finfo)));
+           
             OutputDebugString(_T("没有权限，访问目录！"));
             return -2;
         }
@@ -150,8 +164,8 @@ protected:
 
             FILEINFO finfo;
             finfo.HasNext = FALSE;
-            CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-            CServerSocket::getInstance()->Send(pack);
+
+            listPackets.push_back(CPacket(2, (BYTE*)&finfo, sizeof(finfo)));
 
             return -3;
         }
@@ -161,8 +175,9 @@ protected:
             finfo.IsDirectory = (fdata.attrib & _A_SUBDIR) != 0;
             memcpy(finfo.szFileName, fdata.name, strlen(fdata.name));
             TRACE("[服务器]%s HasNext=%d IsDirectory=%d IsInvalid=%d \r\n", finfo.szFileName, finfo.HasNext, finfo.IsDirectory, finfo.IsInvalid);
-            CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-            CServerSocket::getInstance()->Send(pack);
+             
+            listPackets.push_back(CPacket(2, (BYTE*)&finfo, sizeof(finfo)));
+
             count++;
 
         } while (!_findnext(hfind, &fdata));
@@ -171,42 +186,33 @@ protected:
         //空文件信息，标记结尾。
         FILEINFO finfo;
         finfo.HasNext = FALSE;
-        CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-        if (CServerSocket::getInstance()->Send(pack) == false) {
-            OutputDebugString(_T("发送失败\n"));
-            return -4;
-        }
+        listPackets.push_back(CPacket(2, (BYTE*)&finfo, sizeof(finfo)));
+
         return 0;
     }
 
-    int RunFile() {
-        std::string strPath;
-        if (CServerSocket::getInstance()->GetFilePath(strPath) == false) {
-            OutputDebugString(_T("当前的命令不是启动文件，命令解析错误！"));
-            return -1;
-        }
+    int RunFile(std::list<CPacket>& listPackets, CPacket& inPacket) {
 
+        std::string strPath = inPacket.strData;
+    
         ShellExecuteA(NULL, NULL, strPath.c_str(), NULL, NULL, SW_SHOWNORMAL);
 
-        CPacket pack(3, NULL, 0);
-        if (CServerSocket::getInstance()->Send(pack) == false) {
-            OutputDebugString(_T("发送失败"));
-            return -2;
-        }
+        listPackets.push_back(CPacket(3, NULL, 0));
+
         return 0;
     }
 
     //服务器传给客户端 文件。
-    int DownloadFile() {
-        std::string strPath;
-        CServerSocket::getInstance()->GetFilePath(strPath);
+    int DownloadFile(std::list<CPacket>& listPackets, CPacket& inPacket) {
+        std::string strPath = inPacket.strData;
+
         long long data = 0;     //文件长度
         //FILE* pFile = fopen(strPath.c_str(), "rb");          
         FILE* pFile = NULL;
         errno_t err = fopen_s(&pFile, strPath.c_str(), "rb");   //解决 _WINSOCK_DEPRECATED_NO_WARNINGS 问题
         if (err != 0) {
-            CPacket pack(4, (BYTE*)&data, 0);
-            CServerSocket::getInstance()->Send(pack);
+            listPackets.push_back(CPacket(4, (BYTE*)&data, 0));
+           
             return -1;
         }
 
@@ -214,8 +220,7 @@ protected:
             //发送文件大小信息，用于计算下载文件的进度信息。
             fseek(pFile, 0, SEEK_END);
             data = ftell(pFile);
-            CPacket head(4, (BYTE*)&data, 8);
-            CServerSocket::getInstance()->Send(head);
+            listPackets.push_back(CPacket(4, (BYTE*)&data, 8));
 
             fseek(pFile, 0, SEEK_SET);
 
@@ -223,126 +228,121 @@ protected:
             size_t rlen = 0;
             do {
                 rlen = fread(buffer, 1, 1024, pFile);
-                CPacket pack(4, (BYTE*)&buffer, rlen);
-                CServerSocket::getInstance()->Send(pack);
+
+                listPackets.push_back(CPacket(4, (BYTE*)&buffer, rlen));
+
             } while (rlen >= 1024);
 
             fclose(pFile);
         }
+        listPackets.push_back(CPacket(4, NULL, 0));
 
-        //标记结束
-        CPacket pack(4, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
         return 0;
     }
 
-    int MouseEvent() {
+    int MouseEvent(std::list<CPacket>& listPackets, CPacket& inPacket) {
         MOUSEEV mouse;
-        if (CServerSocket::getInstance()->GetMouseEvent(mouse) == true) {
-            WORD nFlags = 0;
 
-            //确保不在switch里添加if判断语句
-            switch (mouse.nButton)
-            {
-            case 0:     //左键
-                nFlags = 1;
-                break;
-            case 1:     //右键
-                nFlags = 2;
-                break;
-            case 2:     //中键
-                nFlags = 4;
-                break;
-            case 3:     //没有按键
-                nFlags = 8;
-                break;
-            }
+        memcpy(&mouse, inPacket.strData.c_str(), sizeof(MOUSEEV));
 
-            // 鼠标移动用mouse_event
-            SetCursorPos(mouse.ptXY.x, mouse.ptXY.y);
+        WORD nFlags = 0;
 
-
-            switch (mouse.nAction)
-            {
-            case 0: //点击
-                nFlags |= 0x10;
-                break;
-            case 1: //双击
-                nFlags |= 0x20;
-                break;
-            case 2: //按下
-                nFlags |= 0x40;
-                break;
-            case 3: //放开
-                nFlags |= 0x80;
-                break;
-            case 4:
-                break;
-            default:
-                break;
-            }
-
-            switch (nFlags)
-            {
-            case 0x11://左键点击
-                mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x21://左键双击
-                mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x41://左键按下
-                mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x81://左键放开
-                mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-
-            case 0x12://右键点击
-                mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x22://右键双击
-                mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x42://右键按下
-                mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x82://右键放开
-                mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-
-            case 0x14://中键点击
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x24://中键双击
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x44://中键按下
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x84://中键放开
-                mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-
-                //case 0x08://仅鼠标移动
-                //    mouse_event(MOUSEEVENTF_MOVE, mouse.ptXY.x, mouse.ptXY.y, 0, GetMessageExtraInfo());
-                //    break;
-            }
-
-            //结束标记
-            CPacket pack(5, NULL, 0);
-            CServerSocket::getInstance()->Send(pack);
+        //确保不在switch里添加if判断语句
+        switch (mouse.nButton)
+        {
+        case 0:     //左键
+            nFlags = 1;
+            break;
+        case 1:     //右键
+            nFlags = 2;
+            break;
+        case 2:     //中键
+            nFlags = 4;
+            break;
+        case 3:     //没有按键
+            nFlags = 8;
+            break;
         }
-        else {
-            OutputDebugString(_T("获取鼠标操作参数失败"));
-            return -1;
+
+        // 鼠标移动用mouse_event
+        SetCursorPos(mouse.ptXY.x, mouse.ptXY.y);
+
+
+        switch (mouse.nAction)
+        {
+        case 0: //点击
+            nFlags |= 0x10;
+            break;
+        case 1: //双击
+            nFlags |= 0x20;
+            break;
+        case 2: //按下
+            nFlags |= 0x40;
+            break;
+        case 3: //放开
+            nFlags |= 0x80;
+            break;
+        case 4:
+            break;
+        default:
+            break;
         }
+
+        switch (nFlags)
+        {
+        case 0x11://左键点击
+            mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x21://左键双击
+            mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_LEFTDOWN | MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x41://左键按下
+            mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x81://左键放开
+            mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+
+        case 0x12://右键点击
+            mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x22://右键双击
+            mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_RIGHTDOWN | MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x42://右键按下
+            mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x82://右键放开
+            mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+
+        case 0x14://中键点击
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x24://中键双击
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN | MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x44://中键按下
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x84://中键放开
+            mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+
+            //case 0x08://仅鼠标移动
+            //    mouse_event(MOUSEEVENTF_MOVE, mouse.ptXY.x, mouse.ptXY.y, 0, GetMessageExtraInfo());
+            //    break;
+        }
+
+        listPackets.push_back(CPacket(5, NULL, 0));
+
         return 0;
     }
 
-    int SendScreen()
+    int SendScreen(std::list<CPacket>& listPackets, CPacket& inPacket)
     {
         CImage screen;  //GDI
 
@@ -377,8 +377,7 @@ protected:
 
             SIZE_T nSize = GlobalSize(hMem);
 
-            CPacket pack(6, pData, nSize);
-            CServerSocket::getInstance()->Send(pack);
+            listPackets.push_back(CPacket(6, pData, nSize));
 
             GlobalUnlock(hMem);
         }
@@ -391,59 +390,45 @@ protected:
     }
 
     //添加一个 Dialog，样式Popup， 系统菜单False，边框None
-
     //创建线程，方便系统响应解锁消息。
-    int LockMachine()
+    int LockMachine(std::list<CPacket>& listPackets, CPacket& inPacket)
     {
         if (dlg.m_hWnd == NULL || dlg.m_hWnd == INVALID_HANDLE_VALUE) {
-            //_beginthread(threadLockDig, 0, NULL);
             _beginthreadex(NULL, 0, &CCommand::threadEntryForLockDlg, this, 0, &threadId);
             TRACE("[服务器]%s(%d):%d\r\n", __FUNCTION__, __LINE__, threadId);
         }
 
-        CPacket pack(7, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPackets.push_back(CPacket(7, NULL, 0));
 
         return 0;
     }
 
-    int UnlockMachine()
+    int UnlockMachine(std::list<CPacket>& listPackets, CPacket& inPacket)
     {
         //dlg.SendMessage(WM_KEYDOWN, 0x41, 001E0001);
         //::SendMessage(dlg.m_hWnd, WM_KEYDOWN, 0x41, 001E0001);
         //这里是 winapi，从main创建一个线程。需要 向特定的线程 发消息。
         PostThreadMessage(threadId, WM_KEYDOWN, 0x41, 001E0001);
 
-        CPacket pack(8, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPackets.push_back(CPacket(8, NULL, 0));
 
         return 0;
     }
 
-    int DeleteLocalFile()
+    int DeleteLocalFile(std::list<CPacket>& listPackets, CPacket& inPacket)
     {
-        std::string filePath;
-        if (CServerSocket::getInstance()->GetFilePath(filePath) == false)
-        {
-            OutputDebugString(_T("当前的命令不是删除文件，命令解析错误！"));
-            return -1;
-        }
+        std::string filePath = inPacket.strData;
 
         DeleteFile(filePath.c_str());
 
-        //成功后，回应。
-        CPacket pack(9, NULL, 0);
-        if (CServerSocket::getInstance()->Send(pack) == false) {
-            OutputDebugString(_T("发送失败"));
-            return -2;
-        }
+        listPackets.push_back(CPacket(9, NULL, 0));
+
         return 0;
     }
 
-    int TestConnect()
+    int TestConnect(std::list<CPacket>& listPackets, CPacket& inPacket)
     {
-        CPacket pack(2024, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPackets.push_back(CPacket(2024, NULL, 0));
         return 0;
     }
