+++ RemoteCtrl/RemoteCtrl/RemoteCtrl.cpp
@@ -34,35 +34,24 @@ int main()
         }
         else
         {
-            // TODO: 在此处为应用程序的行为编写代码。
-            // 套接字初始化
             CCommand cmd;
             CServerSocket* pserver = CServerSocket::getInstance();
-            if (pserver->InitSocket() == false) {
+            int ret = pserver->Run(&CCommand::RunCommand, &cmd);
+
+            switch (ret) {
+
+            case -1:
                 MessageBox(NULL, _T("网络初始化异常，请检查网络状态！"), _T("网络初始化失败"), MB_OK | MB_ICONERROR);
                 exit(0);
+                break;
+
+            case -2:
+                MessageBox(NULL, _T("多次无法正常接入用户，结束程序！"), _T("接入用户失败"), MB_OK | MB_ICONERROR);
+                exit(0);
+                break;
+
             }
-            int count = 0;
-            while (CServerSocket::getInstance() != NULL) { 
-                if (pserver->AcceptClient() == false) {
-                    if (count >= 3) {
-                        MessageBox(NULL, _T("多次无法正常接入用户，结束程序！"), _T("接入用户失败"), MB_OK | MB_ICONERROR);
-                        exit(0);
-                    }
-                    MessageBox(NULL, _T("无法正常接入用户，自动重试！"), _T("接入用户失败"), MB_OK | MB_ICONERROR);
-                    count++;
-                }
-                int ret = pserver->DealCommand();
-                //TODO
-                if (ret > 0) {
-                    ret = cmd.ExcuteCommand(ret);
-                    if (ret != 0) {
-                        TRACE("[服务器]执行命令失败，%d ret=%d\r\n", pserver->GetPacket().sCmd, ret);
-                    }  
-                    //短连接
-                    pserver->CloseClient();
-                } 
-            }
+
         }
     }
     else
