+++ RemoteCtrl/RemoteCtrl/RemoteCtrl.vcxproj.filters
@@ -42,6 +42,9 @@
     <ClInclude Include="Command.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="Packet.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="RemoteCtrl.cpp">
